requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.1.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud==0.34.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
multidict==6.0.5
oauthlib==3.2.2
openai==1.39.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pyTelegramBotAPI==4.22.0
python-dotenv==1.0.1
python-telegram-bot==21.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
sqlalchemy==2.0.31
pymysql==1.1.1
